Also split electronic into subpackage
